# docs/swagger.yaml
definitions:
  domain.Queue:
    properties:
      id:
        type: integer
      is_open:
        type: boolean
      subject_id:
        type: integer
      title:
        type: string
    required:
    - subject_id
    - title
    type: object
  domain.RefreshTokenInput:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  domain.ResponseUser:
    properties:
      id:
        type: integer
      name:
        type: string
      role:
        type: integer
      username:
        type: string
    type: object
  domain.SignInInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  domain.SignInResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  domain.SignUpInput:
    properties:
      password:
        maxLength: 48
        minLength: 12
        type: string
      register_token:
        type: string
      username:
        type: string
    required:
    - password
    - register_token
    - username
    type: object
  domain.Subject:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  domain.UpdateQueueInput:
    properties:
      is_open:
        type: boolean
      title:
        type: string
    type: object
  domain.UpdateSubjectInput:
    properties:
      name:
        type: string
    required:
    - name
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server celler server.
  title: Group Assistant API
  version: "1.0"
paths:
  /api/auth/jwt/create:
    post:
      operationId: create-jwt
      parameters:
      - description: Credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SignInResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Create JWT
      tags:
      - auth
  /api/auth/jwt/refresh:
    post:
      operationId: refresh-jwt
      parameters:
      - description: Credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.RefreshTokenInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SignInResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Refresh JWT
      tags:
      - auth
  /api/auth/sign-up:
    post:
      description: Sign up. Provide there login, password and registration token
      operationId: sign-up
      parameters:
      - description: Credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.SignUpInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ResponseUser'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Sign up
      tags:
      - auth
  /api/queues:
    post:
      operationId: create-queue
      parameters:
      - description: Queue data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.Queue'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Queue'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create queue
      tags:
      - queues
  /api/queues/{id}:
    delete:
      operationId: delete-queue
      parameters:
      - description: Queue ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Delete queue
      tags:
      - queues
    get:
      operationId: get-queue-by-id
      parameters:
      - description: Queue ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Queue'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Get queue by ID
      tags:
      - queues
    patch:
      description: Allowed to use any field provided in the input body
      operationId: update-queue
      parameters:
      - description: Queue
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateQueueInput'
      - description: Queue ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update queue
      tags:
      - queues
  /api/queues/by_subject/{id}:
    get:
      operationId: get-all-queues
      parameters:
      - description: Subject ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Queue'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get all queues by subject id
      tags:
      - queues
  /api/subjects:
    get:
      operationId: get-all-subjects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Subject'
            type: array
        "500":
          description: Internal Server Error
      summary: Get all subjects
      tags:
      - subjects
    post:
      operationId: create-subject
      parameters:
      - description: Subject data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.Subject'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Subject'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create subject
      tags:
      - subjects
  /api/subjects/{id}:
    get:
      operationId: get-subject-by-id
      parameters:
      - description: Subject ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Subject'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Get subject by ID
      tags:
      - subjects
    put:
      operationId: update-subject
      parameters:
      - description: Subject ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subject data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateSubjectInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update subject
      tags:
      - subjects
  /api/subjects{id}:
    delete:
      operationId: delete-subject
      parameters:
      - description: Subject ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Delete subject
      tags:
      - subjects
securityDefinitions:
  BasicAuth:
    type: basic
  Bearer Token Authentication:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"